Removed icons from buttons to ensure a consistent look
The new menu functions return a button with an SVG icon that slightly
breaks the original default template. Since there's no way to get the
asHtmlButton() function to return a button without the icon, some
substr() trickery was devised to remove it from the returned HTML and
preserve the template's original look.